requirements.txt
Pillow==10.2.0
numpy==1.24.3
scipy==1.11.4
matplotlib==3.8.2